freqt: Creating trees from serialized string rather than a deep copy
Found that creating a tree from a string is faster than a deep copy for
the type of trees used in pl_expand.  Passing a string into pl_expand
also helps make it explicit that the tree is not being expanded in
place, but that a new expanded tree is being created.